requirements.txt
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
chardet==3.0.4
click==6.7
datefinder==0.6.1
decorator==4.2.1
dominate==2.3.1
eatiht==0.1.14
enum34==1.1.6
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
HTMLParser==0.0.2
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
nltk==3.2.5
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
pip-autoremove==0.9.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
recordtype==1.1
regex==2016.1.10
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urlparse2==1.1.1
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.1
